#17: Fixed bug where "Never show this again" didn't work right on first-time disclaimer
If you left that unchecked, meaning you wanted to see it again, then if you moved the window, then next time you start the app, the disclaimer won't show. This is because UserDefaults defaults `Bool`s to `false`. Might try to change the Swift Package I'm using to circumvent this. In the meantime, I just changed it from "should show" to "should skip".